requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
djangorestframework==3.14.0
requests==2.*
geopy==2.4.1
